+++ server/database/manager_getter.py
@@ -3,8 +3,8 @@ from config import settings
 from database.generic_repository import GenericDBManager
 
 def get_db_manager() -> GenericDBManager:
-    from database.sqlite_repository import DevDBManager
-    from database.firestore_repository import FirestoreDBManager
+    from database.sqlite_service import DevDBManager
+    from database.firestore_service import FirestoreDBManager
 
     if settings.env == 'dev' or settings.env == 'test':
         return DevDBManager()
